# swagger.yaml
swagger: '2.0'
info:
  title: Bank Argent API documentation
  description: Contains all available API endpoints in this codebase
  version: '1.0.0'
  termsOfService: 'http://swagger.io/terms/'
host: localhost:8000
basePath: /api/v1/
schemes:
  - http
paths:
  user/transactions:
    post:
      tags:
        - User Module
      summary: transaction
      description: API for transaction
      parameters:
        - in: body
          name: body
          description: Transaction Payload
          required: true
          schema:
            $ref: '#/definitions/Transaction'
      produces:
        - application/json
      responses:
        '200':
          description: transaction came Successfully
          schema:
            type: object
            items:
             type: object
             properties:
              id:
                type: number
                description: the id of the transaction
              date:
                type: string
                description: date of transaction
              description:
                type: string
                description: The name of the company
              amount:
                type: number
                description: The amount of money
              balance:
                type: number
                description: The balance of the account 
        '400':
          description: Invalid Fields
        '500':
          description: Internal Server Error
  /user/transactions:
    get:
      tags:
        - User Module
      summary: get
      description: API for Transaction get 
      produces:
        - application/json
      responses:
        '200':
          description: Get Transaction Successfully
        '400':
          description: Invalid Fields
        '500':
          description: Internal Server Error
 

  /user/transaction/post:
    post:
      security:
        - Bearer: []
      tags:
        - User Module
      summary: Transaction API
      definition:
        ApiResponse:
           type: object
           properties:
             status:
               type: integer
             message:
               type: string
             body:
               type: object
               properties:
                 id:
                   type: string
                 email:
                   type: string
      description: API for fetching a tranaction
      parameters:
        - in: header
          type: string
          name: Authorization
          description: Attach transaction
          required: true
      produces:
        - application/json
      responses:
        '200':
          description: Transaction retrieved successully
          schema:
            $ref: 'ApiResponse'
        '400':
          description: Invalid Fields
        '500':
          description: Internal Server Error
    put:
      security:
        - Bearer: []
      tags:
        - User Module
      summary: Transaction API
      description: API for updating a transaction
      parameters:
        - in: header
          type: string
          name: Authorization
          description: Attach Bearer JWT token
          required: true
        - in: body
          name: body
          description: Update user transaction
          required: true
          schema:
            $ref: '#/definitions/Transaction'
      produces:
        - application/json
      responses:
        '200':
          description: Transaction Data retrieved successully
          schema:
            $ref: 'ApiResponse'
        '400':
          description: Invalid Fields
        '500':
          description: Internal Server Error
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
definitions:
  Transaction:
    properties:
      id:
        type: number
        description: the id of the transaction
      date:
        type: string
        description: date of transaction
      description:
        type: string
        description: The name of the company
      amount:
        type: number
        description: The amount of money
      balance:
        type: number
        description: The balance of the account 
  Additional information:
    properties:
      transaction Type:
        type: string
        description: Electric
      category:
        type: string
        description: Food
      notes:
        type: string
        description: Notes
    